refactor(cross-references): split the annotation passes into sibling modules
Two repo rules the first version missed: `mod.rs` is not allowed, and a new
source file stays within 350 lines. `cross_references/mod.rs` becomes
`cross_references.rs`, and the 602-line `annotate.rs` splits along the seam it
already had — one module per numbering pass, with the shared registry left in
the parent.
